Slug, tailwind réinstall et style refonte
