+++ src/components/ContactForm.astro
@@ -42,7 +42,7 @@
   <div class="text-center">
     <button
       type="submit"
-      class="bg-primary text-secondary-background font-semibold px-6 py-2 rounded hover:bg-opacity-90 transition"
+      class="bg-primary text-secondary-background font-semibold px-6 py-2 rounded transition"
     >
       Envoyer
     </button>
