+++ src/components/Cv.astro
@@ -9,7 +9,7 @@ const cvData = [
     pdf: "/images/pdf/CV Paul KARAS 01:25_bright.pdf",
   },
   {
-    image: "/images/CV Paul KARAS 01:.png",
+    image: "/images/CV Paul KARAS 01:25_dark_photo.png",
     pdf: "/images/pdf/CV Paul KARAS 01:25_dark_photo.pdf",
   },
 ];
