+++ src/components/Footer.astro
@@ -2,11 +2,11 @@
 import Logo from "./Logo.astro";
 ---
 
-<footer class="bg-background/80 text-primary py-8 px-6">
+<footer class="bg-background opacity-80 text-primary py-8 px-6">
   <div
-    class="max-w-7xl mx-auto flex flex-col lg:flex-row justify-between items-center mr-20 gap-8"
+    class="max-w-7xl mx-auto flex flex-col lg:flex-row justify-between items-center gap-8"
   >
-    <div class="text-center lg:text-left text-sm opacity-80">
+    <div class="text-left text-sm opacity-100">
       <p>&copy; 2025 Paul Karas. Tous droits réservés.</p>
       <p>Portfolio réalisé avec Astro et Tailwind CSS.</p>
     </div>
