+++ src/content/config.ts
@@ -1,0 +1,17 @@
+import { defineCollection, z } from "astro:content";
+
+const projects = defineCollection({
+  type: "content",
+  schema: z.object({
+    title: z.string(),
+    image: z.string(),
+    image2: z.string().optional(),
+    image3: z.string().optional(),
+    description: z.string(),
+    link: z.string().optional(),
+  }),
+});
+
+export const collections = {
+  projects,
+};
