+++ src/content/projects/fleurs-et-jardins.md
@@ -1,8 +1,8 @@
 ---
 title: "Fleurs et Jardins"
-slug: "fleurs-et-jardins"
 image: "/images/Salon-code.png"
 description: "Refonte d’un site associatif en Astro pour l’événement de jardinage du Vaudreuil."
+link: "https://github.com/assofleursetjardins/assofleursetjardins.github.io"
 ---
 
 ## Refonte du site du Salon Fleurs & Jardins
