+++ src/content/projects/kart-explorer.md
@@ -1,6 +1,5 @@
 ---
 title: "Kart Explorer"
-slug: "kart-explorer"
 image: "/images/Kart-Explorer.png"
 description: "Compétition alliant Mario Kart en ligne et karting sur piste réelle."
 ---
