+++ src/content/projects/take-kare.md
@@ -1,7 +1,7 @@
 ---
 title: "Take Kare"
-slug: "take-kare"
 image: "/images/Dices.png"
+image2: "/images/Throw_Dices.png"
 description: "Création d’un produit innovant pour la marque Take Kare."
 ---
