+++ src/pages/index.astro
@@ -2,16 +2,6 @@
 import Carrousel from "../components/Carrousel.astro";
 import HomeLayout from "../layouts/HomeLayout.astro";
 import Cv from "../components/Cv.astro";
-
-// layout: "../layouts/HomeLayout.astro";
-// title: "Bienvenue";
-// description: "";
-// image: "src/assets/images/";
-// imgtitle: "";
-// author: "Paul KARAS";
-// date: "06-05-2025";
-// metaTitle: "Page d'accueil";
-// metaDescription: "Bienvenue sur la page d'accueil du Portfolio de Paul KARAS, chef de projets digitaux, développeur web.";
 ---
 
 <HomeLayout title="Accueil">
@@ -32,7 +22,7 @@ import Cv from "../components/Cv.astro";
       <img
         src="/images/Paul-KARAS-Working.png"
         alt="Paul Karas"
-        class="mt-12 w-44 sm:w-40 md:w-76 max-w-full object-cover rounded-xl"
+        class="mt-12 w-44 sm:w-40 md:w-72 max-w-full object-cover rounded-xl"
       />
     </section>
     <Carrousel />
