+++ src/pages/projects.astro
@@ -3,7 +3,9 @@ import { getCollection, type CollectionEntry } from "astro:content";
 import ProjectCard from "../components/ProjectCard.astro";
 import BaseLayout from "../layouts/BaseLayout.astro";
 
-const projects = await getCollection("projects");
+type Project = CollectionEntry<"projects">;
+
+const projects: Project[] = await getCollection("projects");
 ---
 
 <BaseLayout>
@@ -14,9 +16,9 @@ const projects = await getCollection("projects");
 
   <h1 class="text-3xl font-bold text-primary text-center my-8">Mes projets</h1>
 
-  <section class="max-w-5xl mx-auto grid grid-cols-1 sm:grid-cols-2 gap-8 p-4">
+  <section class="max-w-5xl mx-auto grid grid-cols-1 sm:grid-cols-2 gap-22 p-4">
     {
-      projects.map((project) => {
+      projects.map((project: Project) => {
         const { title, image, description } = project.data;
         return (
           <ProjectCard title={title} image={image} description={description} />
