+++ src/pages/projects/[slug].astro
@@ -1,0 +1,42 @@
+---
+import {
+  getEntryBySlug,
+  getCollection,
+  type CollectionEntry,
+} from "astro:content";
+import BaseLayout from "../../layouts/BaseLayout.astro";
+
+type Project = CollectionEntry<"projects">;
+
+export async function getStaticPaths() {
+  const projects = await getCollection("projects");
+  return projects.map((project: Project) => ({
+    params: { slug: project.slug },
+  }));
+}
+
+const { slug } = Astro.params;
+const project = await getEntryBySlug("projects", slug);
+
+if (!project) {
+  throw new Error(`Aucun projet trouvé pour le slug : ${slug}`);
+}
+
+const { title, image, description } = project.data;
+---
+
+<BaseLayout>
+  <article class="max-w-4xl mx-auto py-12 px-4">
+    <h1 class="text-3xl font-bold text-primary mb-6">{title}</h1>
+
+    <img
+      src={image}
+      alt={`Image du projet ${title}`}
+      class="rounded-lg mb-6 w-full object-cover shadow-md"
+    />
+
+    <div class="prose prose-invert max-w-none">
+      {project.body}
+    </div>
+  </article>
+</BaseLayout>
