--- src/pages/projects/interview.astro
@@ -1,11 +1,0 @@
----
-import BaseLayout from "../../layouts/BaseLayout.astro";
----
-
-<BaseLayout>
-  <h1>Interview d'une professionnelle du digital</h1>
-  <h2>Valentine KARAS</h2>
-  <div class="project-images">
-    <img src="../images/Valentine-KARAS.png" alt="Interview Professionnelle" />
-  </div>
-</BaseLayout>
