+++ src/pages/test.astro
@@ -1,0 +1,15 @@
+---
+import "../styles/global.css";
+---
+
+<html lang="fr">
+  <head>
+    <meta charset="UTF-8" />
+    <meta name="viewport" content="width=device-width" />
+    <title>Test pur</title>
+  </head>
+  <body class="bg-white text-black">
+    <h1>Test minimal</h1>
+    <p>Si tu lis ceci, Astro fonctionne.</p>
+  </body>
+</html>
