+++ tsconfig.json
@@ -1,5 +1,8 @@
 {
   "extends": "astro/tsconfigs/strict",
-  "include": [".astro/types.d.ts", "**/*"],
-  "exclude": ["dist"]
+  "include": ["src", "astro.config.mjs"],
+  "exclude": ["dist"],
+  "compilerOptions": {
+    "types": ["astro/client"]
+  }
 }
